% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pws.R
\name{plot_pws}
\alias{plot_pws}
\title{Plotting Station Locations}
\usage{
plot_pws(data_frame)
}
\description{
Plot the locations of the nearest weather stations from the chosen location. Each station is classified as -1, 0, or 1. This will visualize the geographic location and provide information on the quality of data.
' @param data_frame The dataframe from station_lookup_pws().
}
