% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_lookup.R
\name{station_lookup_pws}
\alias{station_lookup_pws}
\title{Station Lookup}
\usage{
station_lookup_pws(lat, lon, apikey)
}
\arguments{
\item{lat}{The latitude in decimal degrees}

\item{lon}{The longitude in decimal degrees}

\item{apikey}{Your Wunderground API key. This can be set ahead of time as an object.}
}
\description{
Look up the 10 closest weather stations to the location entered
}
