% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_grab.R
\name{weather_grab}
\alias{weather_grab}
\title{Single day weather grab}
\usage{
weather_grab(stationID, date, apikey)
}
\arguments{
\item{stationID}{The station ID you want to obtain the weather data.}

\item{date}{A single day to obtain weather data. In the format 'yyyy-mm-dd'. e.g., '2021-09-17'}

\item{apikey}{Your APIKey from wunderground.}
}
\description{
Collect weather data for a single day from a single station.
}
