% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_grab_range.R
\name{weather_grab_range}
\alias{weather_grab_range}
\title{Weather scrape from range of dates}
\usage{
weather_grab_range(stationID, start_date, end_date, apikey)
}
\arguments{
\item{stationID}{The station ID you want to obtain the weather data.}

\item{start_date}{The first day you want to start collecting data. In the format yyyy-mm-dd.}

\item{end_date}{The last day you want to start collecting data. In the format yyyy-mm-dd.}

\item{apikey}{Your APIKey from wunderground.}
}
\description{
Collect data from a station between two dates. If the station was not active during the requested dates, you will receive a warning message.
}
